# food_db/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Macro_vision",
        "title": "Macro_vision backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/meal-cards": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Retrieve a meal card and its dishes belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Get meal card with dishes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal card ID (UUID)",
                        "name": "card_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mealcards.MealCardDish"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Meal card not found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Update meal card information belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Update meal card",
                "parameters": [
                    {
                        "description": "Updated meal card payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mealcards.UpdateMealCardParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mealcards.MealCardDish"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Meal card not found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new meal card for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Create meal card",
                "parameters": [
                    {
                        "description": "Meal card payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mealcards.CreateMealCardParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "new card uuid",
                        "schema": {
                            "$ref": "#/definitions/mealcards.MealCardDishID"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Remove a meal card belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Remove meal card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal card ID (UUID)",
                        "name": "card_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RemoveMealCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Meal card not found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/meal-cards/daily": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Retrieve all meal cards for a specific day belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Get daily meal cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.MealCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "No meal cards found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/meal-cards/dishes": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Add a dish to a meal card belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Add dish to meal card",
                "parameters": [
                    {
                        "description": "Add dish to meal card payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mealcards.AddDishToCardParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AddDishToCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Meal card or dish not found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Remove a dish from a meal card belonging to the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Remove dish from meal card",
                "parameters": [
                    {
                        "description": "Remove dish from meal card payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mealcards.RemoveDishFromCardParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RemoveDishFromCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Meal card or dish not found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/meal-cards/monthly": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Retrieve all meal cards for the month of the given date belonging to the authenticated user. The day portion of the date is ignored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal-cards"
                ],
                "summary": "Get monthly meal cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference date (YYYY-MM-DD); day is ignored",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.MealCard"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "No meal cards found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Dish": {
            "type": "object",
            "properties": {
                "alt_name": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "course": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dish_id": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "full_recipe": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "database.MealCard": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "meal_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "handlers.AddDishToCardResponse": {
            "type": "object",
            "properties": {
                "added_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RemoveDishFromCardResponse": {
            "type": "object",
            "properties": {
                "removed_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RemoveMealCardResponse": {
            "type": "object",
            "properties": {
                "removed_id": {
                    "type": "string"
                }
            }
        },
...
        "mealcards.AddDishToCardParam": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "string"
                },
                "dish_id": {
                    "type": "string"
                }
            }
        },
        "mealcards.CreateMealCardParam": {
            "type": "object",
            "properties": {
                "dishes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meal_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "mealcards.MealCardDish": {
            "type": "object",
            "properties": {
                "dishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Dish"
                    }
                },
                "mealCard": {
                    "$ref": "#/definitions/database.MealCard"
                }
            }
        },
        "mealcards.MealCardDishID": {
            "type": "object",
            "properties": {
                "dishesID": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mealCard": {
                    "$ref": "#/definitions/database.MealCard"
                }
            }
        },
        "mealcards.RemoveDishFromCardParam": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "string"
                },
                "dish_id": {
                    "type": "string"
                }
            }
        },
        "mealcards.UpdateMealCardParam": {
            "type": "object",
            "properties": {
                "meal_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}